R6 object for imported data via the EA API

This is the base class for all data imported via the API. A HydroImport contains public raw data and private metadata.

Value

A class containing raw data, metadata and methods

Public fields

data

Imported data via the API tool. Uses data.table.

Methods

Public methods

HydroImportFactory$new()

HydroImportFactory$print()

HydroImportFactory$methods()

HydroImportFactory$summary()

HydroImportFactory$asVol()

HydroImportFactory$hydroYearDay()

HydroImportFactory$rmVol()

HydroImportFactory$rmHY()

HydroImportFactory$rmHYD()

HydroImportFactory$coords()

HydroImportFactory$nrfa()

HydroImportFactory$dataAgg()

HydroImportFactory$rollingAggs()

HydroImportFactory$meta()

HydroImportFactory$dayStats()

HydroImportFactory$flowDuration()

HydroImportFactory$quality()

HydroImportFactory$clone()

Method new()

Initialise the new HydroImport object

Usage

HydroImportFactory$new(
  data = NA,
  dataType = "Raw Import",
  modifications = NA,
  stationName = NA,
  riverName = NA,
  WISKI = NA,
  RLOID = NA,
  stationGuide = NA,
  baseURL = NA,
  dataURL = NA,
  measureURL = NA,
  idNRFA = NA,
  urlNRFA = NA,
  easting = NA,
  northing = NA,
  latitude = NA,
  longitude = NA,
  area = NA,
  parameter = NA,
  unitName = NA,
  unit = NA,
  datum = NA,
  boreholeDepth = NA,
  aquifer = NA,
  start = NA,
  end = NA,
  timeStep = NA,
  timeZone = "GMT - UTC±00:00",
  records = NA
)

Arguments

data

Raw data

dataType

Details the type of data in this environment

modifications

Details modifications made to the data

stationName

Name of gauge

riverName

River name

WISKI

WISKI ID

RLOID

River Levels on the Internet ID

stationGuide

Station Unique Identifier

baseURL

Base URL used in the API

dataURL

End of URL that downloaded the raw data

measureURL

Primarily used in gaining the metadata

idNRFA

National River Flow Archive station identifier

urlNRFA

National River Flow Archive station URL

easting

Easting coordinate

northing

Northing coordinate

latitude

Latitude coordinate

longitude

Longitude coordinate

area

Catchment area of flow/level gauge

parameter

Details the collected data e.g. flow, level etc.

unitName

Unit name used

unit

Details on the unit of measurement

datum

Datum of the gauge, if available

boreholeDepth

Depth of borehole recording

aquifer

Details on the aquifer type

start

Function calculates the date of the first imported record

end

Function calculates the date of the last imported record

timeStep

Function that calculates the timestep in seconds

timeZone

Time zone used in data, defaults to GMT

records

Function calculates the number of records

Method print()

Display the R6 object

Usage

HydroImportFactory$print(.)

Arguments

.

(ignored).

Method methods()

Display the methods available in the R6 object

Usage

HydroImportFactory$methods(.)

Arguments

.

(ignored).

Method summary()

Display a summary of the R6 object

Usage

HydroImportFactory$summary(.)

Arguments

.

(ignored).

Method asVol()

Calculate the volume of water for flow or rainfall data

Usage

HydroImportFactory$asVol(.)

Arguments

.

(ignored).

Method hydroYearDay()

Calculate hydrological year and day

Usage

HydroImportFactory$hydroYearDay(.)

Arguments

.

(ignored).

Method rmVol()

Remove the calculated volume

Usage

HydroImportFactory$rmVol(.)

Arguments

.

(ignored).

Method rmHY()

Remove the calculated hydroYear

Usage

HydroImportFactory$rmHY(.)

Arguments

.

(ignored).

Method rmHYD()

Remove the calculated hydrological day

Usage

HydroImportFactory$rmHYD(.)

Arguments

.

(ignored).

Method coords()

Return the coordinates of the gauge

Usage

HydroImportFactory$coords(.)

Arguments

.

(ignored).

Method nrfa()

Return the NRFA details of the gauge

Usage

HydroImportFactory$nrfa(.)

Arguments

.

(ignored).

Method dataAgg()

Return aggregated data

Usage

HydroImportFactory$dataAgg(type = NULL, method = NULL)

Arguments

type

Set the aggrgegation level to either hourly', 'daily', 'monthly', 'annual', or 'hydroYear'

method

Available aggregation methods include 'min', 'max', 'mean', 'median', and 'sum'

Method rollingAggs()

Return the different user selecting rolling aggregations

Usage

HydroImportFactory$rollingAggs(
  method = "mean",
  rolls = c(1, 2, 3, 4, 8, 24, 120),
  interval = 0.25
)

Arguments

method

Choose mean, median, min, max, and sum for volumes

rolls

The hourly aggregations selected by the user

interval

Corrects for the time step of the data, set to 0.25 hours

Method meta()

Returns the metadata as a data.table

Usage

HydroImportFactory$meta(.)

Arguments

.

(ignored).

Method dayStats()

Daily statistics for imported data

Usage

HydroImportFactory$dayStats(plot = TRUE, year = NULL)

Arguments

plot

Set to TRUE, will produce a plot of the statistics

year

Set to NULL, overlays the mean daily flow for a chosen hydrological year

Method flowDuration()

Flow duration curve of data

Usage

HydroImportFactory$flowDuration(perc = c(5, 25, 75, 95))

Arguments

perc

Determine flow percentiles. Set to c(5, 25, 75, 95)

Method quality()

Displays the number of observations under each quality flag

Usage

HydroImportFactory$quality(.)

Arguments

.

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

HydroImportFactory$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.