annualAgg

Aggregates sub annual time series into an annual resolution, aggregations are carried out to the calendar day not a rolling time period.

Usage

annualAgg(x, method = "mean", ...)

# S3 method for class 'data.table'
annualAgg(x, method = "mean", ...)

# S3 method for class 'HydroImport'
annualAgg(x, method = "mean", ...)

Arguments

x

Data generated in the riskyData package

method

'mean', 'median', 'max', 'min', or 'sum'

...

Other variables as required

Value

A dataset of annually aggregated data, data are not stored in a HydroAggs container.

Examples

data(bewdley)

# Calculate the mean annual (calendar year) flows
annualAgg(bewdley)
#>     dateTime annualMean
#>        <int>      <num>
#>  1:     2008   89.75748
#>  2:     2009   52.69865
#>  3:     2010   39.04404
#>  4:     2011   39.62156
#>  5:     2012   78.88227
#>  6:     2013   65.46530
#>  7:     2014   76.83255
#>  8:     2015   59.09761
#>  9:     2016   61.14972
#> 10:     2017   48.04278
#> 11:     2018   60.52330
#> 12:     2019   78.14450
#> 13:     2020   80.21779
#> 14:     2021   63.70356
#> 15:     2022   44.42093