Find the amount of events over a given threshold and supply the amount of time steps to which they were in excess.

Find the amount of events over a given threshold and supply the amount of time steps to which they were in excess.

Usage

exceed(
  dateTime = NULL,
  value = NULL,
  threshold = NULL,
  gapWidth = 0,
  timeStep = 900
)

Arguments

dateTime

The date and time series

value

The observed data

threshold

The threshold at which to test

gapWidth

This allows you to ignore a set number of time steps beneath a threshold between two or more events

timeStep

The time step used applied in the observed data in seconds

Value

A table of events with start and end times

Examples

data(bewdley)
events <- exceed(bewdley$data$dateTime,
                 bewdley$data$value,
                 threshold = 200,
                 gapWidth = 20,
                 timeStep = 900)
events
#>                   Start                 End timeSteps
#>                  <POSc>              <POSc>     <num>
#>  1: 2008-10-07 20:30:00 2008-10-08 22:00:00       102
#>  2: 2008-11-11 12:00:00 2008-11-11 12:15:00         1
#>  3: 2008-11-11 18:45:00 2008-11-11 22:45:00        16
#>  4: 2008-11-12 12:45:00 2008-11-12 13:00:00         1
#>  5: 2009-11-20 03:00:00 2009-12-02 23:30:00      1234
#>  6: 2010-01-17 07:30:00 2010-01-20 12:30:00       308
#>  7: 2011-01-14 12:15:00 2011-01-16 10:00:00       183
#>  8: 2011-01-16 15:15:00 2011-01-16 15:30:00         1
#>  9: 2011-01-17 07:00:00 2011-01-19 07:00:00       192
#> 10: 2011-02-07 13:45:00 2011-02-10 17:45:00       304
#> 11: 2012-01-05 06:30:00 2012-01-09 00:45:00       361
#> 12: 2012-04-28 03:30:00 2012-04-29 05:00:00       102
#> 13: 2012-04-29 11:30:00 2012-05-03 02:00:00       346
#> 14: 2012-07-14 03:00:00 2012-07-14 07:00:00        16
#> 15: 2012-07-15 14:00:00 2012-07-15 19:30:00        22
#> 16: 2012-09-25 17:00:00 2012-09-28 23:00:00       312
#> 17: 2012-11-24 18:45:00 2012-11-30 22:15:00       590
#> 18: 2012-12-08 19:30:00 2012-12-09 00:45:00        21
#> 19: 2012-12-16 23:15:00 2012-12-17 16:00:00        67
#> 20: 2012-12-20 06:00:00 2013-01-05 11:45:00      1559
#> 21: 2013-01-27 19:15:00 2013-02-03 13:15:00       648
#> 22: 2013-02-11 17:45:00 2013-02-13 03:15:00       134
#> 23: 2013-02-15 01:00:00 2013-02-17 10:15:00       229
#> 24: 2013-05-16 16:45:00 2013-05-18 04:30:00       143
#> 25: 2013-12-23 04:00:00 2013-12-29 14:45:00       619
#> 26: 2014-01-01 14:45:00 2014-01-13 09:45:00      1132
#> 27: 2014-01-20 10:00:00 2014-01-21 04:15:00        73
#> 28: 2014-01-27 23:00:00 2014-02-21 04:15:00      2325
#> 29: 2014-12-12 15:00:00 2014-12-16 15:00:00       384
#> 30: 2014-12-20 16:30:00 2014-12-22 07:15:00       155
#> 31: 2015-01-17 00:30:00 2015-01-19 03:15:00       203
#> 32: 2015-12-02 20:00:00 2015-12-07 18:15:00       473
#> 33: 2015-12-14 16:15:00 2015-12-18 05:00:00       339
#> 34: 2015-12-18 10:45:00 2015-12-18 11:00:00         1
#> 35: 2015-12-24 23:00:00 2015-12-26 21:15:00       185
#> 36: 2015-12-27 03:30:00 2015-12-31 08:45:00       405
#> 37: 2016-01-01 10:45:00 2016-01-01 11:00:00         1
#> 38: 2016-01-01 16:45:00 2016-01-06 11:15:00       458
#> 39: 2016-01-09 15:15:00 2016-01-16 04:00:00       627
#> 40: 2016-02-08 14:15:00 2016-02-13 17:00:00       491
#> 41: 2017-03-04 21:45:00 2017-03-04 22:00:00         1
#> 42: 2017-12-15 11:45:00 2017-12-18 05:00:00       261
#> 43: 2018-01-05 02:00:00 2018-01-05 06:45:00        19
#> 44: 2018-01-05 18:00:00 2018-01-07 00:00:00       120
#> 45: 2018-01-23 05:30:00 2018-01-27 18:45:00       437
#> 46: 2018-04-03 19:15:00 2018-04-06 23:00:00       303
#> 47: 2018-12-22 21:15:00 2018-12-25 20:15:00       284
#> 48: 2019-03-16 16:00:00 2019-03-21 06:45:00       443
#> 49: 2019-06-13 08:45:00 2019-06-17 22:45:00       440
#> 50: 2019-10-01 15:30:00 2019-10-05 12:45:00       373
#> 51: 2019-10-26 10:45:00 2019-10-31 22:30:00       527
#> 52: 2019-11-09 18:15:00 2019-11-11 03:15:00       132
#> 53: 2019-11-13 07:15:00 2019-11-19 21:45:00       634
#> 54: 2019-12-13 19:45:00 2019-12-23 07:15:00       910
#> 55: 2020-01-14 19:15:00 2020-01-17 09:45:00       250
#> 56: 2020-02-10 20:30:00 2020-03-06 05:30:00      2340
#> 57: 2020-03-11 20:30:00 2020-03-14 10:15:00       247
#> 58: 2020-12-18 19:15:00 2021-01-01 10:00:00      1307
#> 59: 2021-01-20 15:15:00 2021-01-27 04:45:00       630
#> 60: 2021-01-28 03:30:00 2021-02-07 10:45:00       989
#> 61: 2021-02-22 10:15:00 2021-02-24 06:45:00       178
#> 62: 2021-03-15 00:00:00 2021-03-16 16:00:00       160
#> 63: 2021-12-10 15:45:00 2021-12-13 04:00:00       241
#> 64: 2022-02-17 07:15:00 2022-02-27 06:15:00       956
#>                   Start                 End timeSteps