Download hydrometric data from the Hydrology Data Explorer

Used in the loadAPI() function

Usage

hdeAPI(
  ID = NULL,
  measure = NULL,
  period = NULL,
  type = NULL,
  datapoints = "standard",
  from = NULL,
  to = NULL,
  rtExt = FALSE,
  meta = TRUE,
  rtLookup = FALSE
)

Arguments

ID

Use to specify a particular WISKI ID or for when downloading all stations 'wiski' or 'nrfa' to filter results. Additionally to access the real time APIs use "flood" for all critical sites and "tidal" for the category A tide gauges.

measure

Use this when exporting observations to select the available parameter. Generally 'flow', 'level', or 'groundwater'

period

This is so you can select the time steps available, generally 900 (15 min) or 86400 (daily)

type

Selects the data type, instantaneous, min, max etc.

datapoints

Allows you to decide where in the time series you are pulling the data from

from

First time step of choice

to

Last time step of choice

rtExt

Set to NULL, if TRUE hydrological data series are extended with the real time API

meta

Set as TRUE, exports metadata for gauge with the data request

rtLookup

Set as FALSE, this uses an inbuilt dataset, if set to TRUE this will update the table using the most recent data. However, this can dramatically increase the run time.

Value

r6 HydroImport container

Examples

#hdeAPI(ID = '2001', measure = 'flow', period = 900)