hydroYearAgg

Aggregates sub annual time series into a hydrological year resolution, aggregations are carried out to the calendar day not a rolling time period.

Usage

hydroYearAgg(x, method = "mean", ...)

# S3 method for class 'data.table'
hydroYearAgg(x, method = "mean", ...)

# S3 method for class 'HydroImport'
hydroYearAgg(x, method = "mean", ...)

Arguments

x

Data generated in the riskyData package

method

'mean', 'median', 'max', 'min', or 'sum'

...

Other variables as required

Value

Data are aggregated to the hydrological year, data are not stored in a HydroAggs container.

Examples

data(bewdley)

# Calculate the hydrological year
bewdley$hydroYearDay()
#> ℹ Calculating hydrological year and day
#> ✔ Calculating hydrological year and day [2.3s]
#> 

# Calculate the annual maximum flows for Bewdley (hydrological year)
hydroYearAgg(bewdley, method = "max")
#>     hydroYear hydroYearMax
#>         <num>        <num>
#>  1:      2009          213
#>  2:      2010          312
#>  3:      2011          346
#>  4:      2012          266
#>  5:      2013          376
#>  6:      2014          443
#>  7:      2015          250
#>  8:      2016          323
#>  9:      2017          208
#> 10:      2018          255
#> 11:      2019          291
#> 12:      2020          523
#> 13:      2021          496
#> 14:      2022          470