A means to pair up hydrological events by nearest match time

A means to pair up hydrological events by nearest match time

Usage

linkEvents(
  upstream = NULL,
  downstream = NULL,
  timeMin = 0,
  timeMax = 1000,
  matrix = TRUE
)

Arguments

upstream

Peak values and date time of upstream site

downstream

Peak values and date time of downstream site

timeMin

Used to filter the time difference in paired peaks, set as 0

timeMax

Used to filter the time difference in paired peaks, set as 1000

matrix

Set to true, exports a matrix of difftimes in minutes that can be used to highlight the peak matching. If set as FALSE a nearest join is used.

Value

a data.table of paired events

Examples

bewdS <- loadAPI(ID = '2001', 'level', 900, 'instantaneous',
                 datapoints = 'range',
                 from = '2008-10-01 09:00')
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [911ms]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [25.8s]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [1.2s]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [12ms]
#> 
                 bewdS$hydroYearDay()
#> ℹ Calculating hydrological year and day
#> ✔ Calculating hydrological year and day [2.3s]
#> 

buildS <- loadAPI(ID = '2134', 'level', 900, 'instantaneous',
                  datapoints = 'range',
                  from = '2008-10-01 09:00')
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [858ms]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [24.7s]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [1s]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [12ms]
#> 
buildS$hydroYearDay()
#> ℹ Calculating hydrological year and day
#> ✔ Calculating hydrological year and day [2.6s]
#> 

bewdPS <- findPeaks(bewdS, levels = 100)
buildPS <- findPeaks(buildS, levels = 100)
pairedStage <- linkEvents(upstream = buildPS,
                          downstream = bewdPS)
pairedStage
#>      upstreamPos downstreamPos
#>           <list>         <int>
#>   1:           1             1
#>   2:           2             2
#>   3:           3             3
#>   4:           4             4
#>   5:           5             5
#>  ---                          
#> 771:         671           771
#> 772:         672           772
#> 773:                       773
#> 774:         674           774
#> 775:                       775

## Plot the paired events
plot(pairedStage$value ~ pairedStage$`i.value`,
     xlab = 'Buildwas Stage (m)',
     ylab = "Bewdley Stage (m)")
#> Error in (function (formula, data = NULL, subset = NULL, na.action, drop.unused.levels = FALSE,     xlev = NULL, ...) {    possible_newdata <- !missing(data) && is.data.frame(data) &&         identical(substitute(data), quote(newdata)) && (nr <- nrow(data)) >         0    if (!missing(formula) && nargs() == 1 && is.list(formula) &&         !is.null(m <- formula$model))         return(m)    if (!missing(formula) && nargs() == 1 && is.list(formula) &&         all(c("terms", "call") %in% names(formula))) {        fcall <- formula$call        m <- match(c("formula", "data", "subset", "weights",             "na.action"), names(fcall), 0)        fcall <- fcall[c(1, m)]        fcall[[1L]] <- quote(stats::model.frame)        env <- environment(formula$terms) %||% parent.frame()        return(eval(fcall, env))    }    if (missing(formula)) {        if (!missing(data) && inherits(data, "data.frame") &&             length(attr(data, "terms")))             return(data)        formula <- as.formula(data)    }    else if (missing(data) && inherits(formula, "data.frame")) {        if (length(attr(formula, "terms")))             return(formula)        data <- formula        formula <- as.formula(data)    }    else formula <- as.formula(formula)    if (missing(na.action)) {        na.action <- if (!is.null(naa <- attr(data, "na.action")) &&             mode(naa) != "numeric")             naa        else getOption("na.action") %||% na.fail    }    if (missing(data))         data <- environment(formula)    else if (!is.data.frame(data) && !is.environment(data) &&         !is.null(attr(data, "class")))         data <- as.data.frame(data)    else if (is.array(data))         stop("'data' must be a data.frame, not a matrix or an array")    if (!is.data.frame(data) && !is.environment(data) && !is.list(data) &&         !is.null(data))         stop("'data' must be a data.frame, environment, or list")    if (!inherits(formula, "terms"))         formula <- terms(formula, data = data)    env <- environment(formula)    rownames <- .row_names_info(data, 0L)    vars <- attr(formula, "variables")    predvars <- attr(formula, "predvars") %||% vars    varnames <- vapply(vars, deparse2, " ")[-1L]    variables <- eval(predvars, data, env)    resp <- attr(formula, "response")    if (is.null(rownames) && resp > 0L) {        lhs <- variables[[resp]]        rownames <- if (is.matrix(lhs))             rownames(lhs)        else names(lhs)    }    if (possible_newdata && length(variables)) {        nr2 <- max(sapply(variables, NROW))        if (nr2 != nr)             warning(sprintf(paste0(ngettext(nr, "'newdata' had %d row",                 "'newdata' had %d rows"), " ", ngettext(nr2,                 "but variable found had %d row", "but variables found have %d rows")),                 nr, nr2), call. = FALSE, domain = NA)    }    if (is.null(attr(formula, "predvars"))) {        for (i in seq_along(varnames)) predvars[[i + 1L]] <- makepredictcall(variables[[i]],             vars[[i + 1L]])        attr(formula, "predvars") <- predvars    }    extras <- substitute(list(...))    extranames <- names(extras[-1L])    extras <- eval(extras, data, env)    subset <- eval(substitute(subset), data, env)    data <- .External2(C_modelframe, formula, rownames, variables,         varnames, extras, extranames, subset, na.action)    if (length(xlev)) {        for (nm in names(xlev)) if (!is.null(xl <- xlev[[nm]])) {            xi <- data[[nm]]            if (is.character(xi))                 xi <- as.factor(xi)            if (!is.factor(xi) || is.null(nxl <- levels(xi)))                 warning(gettextf("variable '%s' is not a factor",                   nm), domain = NA)            else {                ctr <- attr(xi, "contrasts")                xi <- xi[, drop = TRUE]                nxl <- levels(xi)                if (any(m <- is.na(match(nxl, xl))))                   stop(sprintf(ngettext(length(m), "factor %s has new level %s",                     "factor %s has new levels %s"), nm, paste(nxl[m],                     collapse = ", ")), domain = NA)                data[[nm]] <- factor(xi, levels = xl, exclude = NULL)                if (!identical(attr(data[[nm]], "contrasts"),                   ctr))                   warning(gettextf("contrasts dropped from factor %s",                     nm), call. = FALSE, domain = NA)            }        }    }    else if (drop.unused.levels) {        for (nm in names(data)) {            x <- data[[nm]]            if (is.factor(x) && length(unique(x[!is.na(x)])) <                 length(levels(x))) {                ctr <- attr(x, "contrasts")                data[[nm]] <- x[, drop = TRUE]                if (!identical(attr(data[[nm]], "contrasts"),                   ctr))                   warning(gettextf("contrasts dropped from factor %s due to missing levels",                     nm), domain = NA, call. = FALSE)            }        }    }    attr(formula, "dataClasses") <- vapply(data, .MFclass, "")    attr(data, "terms") <- formula    data})(formula = pairedStage$value ~ pairedStage$i.value, na.action = NULL): invalid type (NULL) for variable 'pairedStage$value'
model <- lm(pairedStage$value ~ pairedStage$`i.value`)
#> Error in model.frame.default(formula = pairedStage$value ~ pairedStage$i.value,     drop.unused.levels = TRUE): invalid type (NULL) for variable 'pairedStage$value'
abline(model, lwd = 2, col = 2)
#> Error in eval(expr, envir, enclos): object 'model' not found