Merge HydroImport and HydroAggs R6 objects

The mergeData() is used to combine multiple R6 encapsulations. The data have to be in HydroImport or HydroAggs classes. Depending on the metadata argument, metadata can be exported with the merged dataset as a list. The merging process is carried out on looped full joins, this allows datasets f varying lengths to be collated. Missing information is padded with NAs.

Usage

mergeData(..., metadata = FALSE)

Arguments

...

Datasets downloaded from the EAs API set inside an R6 container

metadata

Set to FALSE, if TRUE collated metadata are exported with the merged dataset

Value

A merged dataset as a data,table, if metadata is TRUE then a list of 2 data.tables

Examples

a <- loadAPI(ID = 2001,
             measure = 'flow',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-12 09:00',
             to = '2023-06-13 09:00')
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [2.6s]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [1.5s]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [2.2s]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [13ms]
#> 

b <- loadAPI(ID = 2002,
             measure = 'flow',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-12 09:00',
             to = '2023-07-12 09:00')
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [2.6s]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [2.1s]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [2.1s]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [13ms]
#> 

c <- loadAPI(ID = 2004,
             measure = 'flow',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-13 09:00',
             to = '2023-07-12 09:00')
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [2.6s]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [2s]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [2.2s]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [13ms]
#> 

d <- loadAPI(ID = 2001,
             measure = 'level',
             period = 900,
             type = 'instantaneous',
             datapoints = 'range',
             from = '2023-06-12 09:00',
             to = '2023-07-12 09:00')
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [2.8s]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [2s]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [1.9s]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [13ms]
#> 

z <- mergeData(a, b, c, d, metadata = FALSE)
#> Warning: Different parametrs will be collated, this could lead to problems in other analyses
z
#> Key: <dateTime>
#>                  dateTime a_value b_value c_value d_value
#>                    <POSc>   <num>   <num>   <num>   <num>
#>    1: 2023-06-12 09:00:00   15.44  12.260      NA   0.341
#>    2: 2023-06-12 09:15:00   15.09  12.606      NA   0.338
#>    3: 2023-06-12 09:30:00   15.15  12.408      NA   0.336
#>    4: 2023-06-12 09:45:00   15.13  12.507      NA   0.334
#>    5: 2023-06-12 10:00:00   14.78  12.507      NA   0.334
#>   ---                                                    
#> 2877: 2023-07-12 08:00:00      NA  11.420   3.699   0.306
#> 2878: 2023-07-12 08:15:00      NA  11.567   3.756   0.308
#> 2879: 2023-07-12 08:30:00      NA  11.665   3.872   0.307
#> 2880: 2023-07-12 08:45:00      NA  11.468   3.950   0.310
#> 2881: 2023-07-12 09:00:00      NA  11.566   4.029   0.308