R6 object for imported data via the EA API

This is the base class for all data imported via the API. A HydroAggs contains public raw data and private metadata.

Value

A class containing raw data, metadata and methods

Super class

riskyData::HydroImport -> HydroAggs

Methods

Public methods

HydroAggsFactory$print()

HydroAggsFactory$clone()

Inherited methods

riskyData::HydroImport$addRating()

riskyData::HydroImport$asVol()

riskyData::HydroImport$coords()

riskyData::HydroImport$dataAgg()

riskyData::HydroImport$dayStats()

riskyData::HydroImport$exceed()

riskyData::HydroImport$findPeaks()

riskyData::HydroImport$flowDuration()

riskyData::HydroImport$hydroYearDay()

riskyData::HydroImport$initialize()

riskyData::HydroImport$meta()

riskyData::HydroImport$methods()

riskyData::HydroImport$missing()

riskyData::HydroImport$nrfa()

riskyData::HydroImport$plot()

riskyData::HydroImport$quality()

riskyData::HydroImport$rateFlow()

riskyData::HydroImport$rateStage()

riskyData::HydroImport$rmHY()

riskyData::HydroImport$rmHYD()

riskyData::HydroImport$rmVol()

riskyData::HydroImport$rollingAggs()

riskyData::HydroImport$summary()

riskyData::HydroImport$window()

Method print()

Display the R6 object

Usage

HydroAggsFactory$print(.)

Arguments

.

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

HydroAggsFactory$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.