DEPRECIATED Load multiple HydroImport containers

The loadMulti() function iterates through combinations of parameters supplied to the arguments. These can be added as vectors, this reduces the amount of calls to loadAPI() in working scripts. Objects or HydroImport are automatically assigned to the environment using a name convention of paramater_WISKI-ID.

Usage

loadMulti(
  API = "EA",
  ID = NULL,
  measure = NULL,
  period = NULL,
  type = NULL,
  datapoints = NULL,
  from = NULL,
  to = NULL,
  ...
)

Arguments

API

Currently only supports the Environment Agency API set as "EA".

ID

Use to specify a particular WISKI ID or for when downloading all stations 'wiski' or 'nrfa' to filter results. Additionally to access the real time APIs use "flood" for all critical sites and "tidal" for the category A tide gauges. Vectors are supported.

measure

Use this when exporting observations to select the available parameter. Generally 'flow', 'level', or 'groundwater'

period

This is so you can select the time steps available, generally 900 (15 min) or 86400 (daily).

type

Selects the data type, instantaneous, min, max etc.

datapoints

Allows you to decide where in the time series you are pulling the data from.

from

First time step of choice.

to

Last time step of choice.

...

Additional loadAPI() arguments can be called.

Value

Multiple HydroImport containers

Examples

loadMulti(ID = c('2001', '2002'),
          measure = c('flow', 'level'),
          period = 900,
          type = 'instantaneous',
          datapoints = 'earliest')
#> ℹ  File names are in the format of measure_ID format
#> 
#> ── Data import: flow_2001 
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [844ms]
#> 
#> ℹ Downloading raw data
#> ✔ Downloading raw data [415ms]
#> 
#> ℹ Collating metadata
#> ✔ Collating metadata [911ms]
#> 
#> ℹ Exporting data to HydroImport container
#> ✔ Exporting data to HydroImport container [12ms]
#> 
#> ℹ  Storing data in environment as flow_2001
#> 
#> ── Data import: level_2002 
#> ℹ Compiling parameters for raw download
#> ✔ Compiling parameters for raw download [842ms]
#> 
#> ℹ Downloading raw data
#> Warning: cannot open URL 'http://environment.data.gov.uk/hydrology/id/measures//readings.json?earliest': HTTP status was '404 Not Found'
#> ✖ Downloading raw data [414ms]
#> 
#> ✖ Failed to download data for level_2002
#> ℹ  Try using the loadAPI() function instead
#> 
#> ──  Errors detected at the following sites: 
#> 2002